requirements.txt
discord.py==2.6.0
python-dotenv>=1.0.0
aiohttp>=3.8.1
PyNaCl>=1.5.0
yt-dlp>=2023.1.0
